+++ changelog.markdown
@@ -88,6 +88,11 @@ Bear in mind that I didn't exactly start this until recently, and FHSTwitterEngi
 - Replaced -getFriends and -getFollowers with the methods for the above
 - Exposed -generateRequestStringsFromArray:
 
+**Version 1.6.3.1**
+- Found an issue with the mass listing method - cursors
+- Adapted code for cursors, no more -getFriends and -getFollowers. You'll just have to pass your logged in username… Boo hoo.
+
+
 
 
 OAuthConsumer